oscar/requirements.txt
appmap>=1.1.0.dev0
# These requirements are only necessary when developing on Oscar.

# development
django-debug-toolbar>=2.2,<4.3
django-extensions>=2.2,<3.3
psycopg2-binary>=2.8,<2.10
django~=4.2.0
pytest~=7.4.2

# Sandbox
Whoosh>=2.7,<2.8
django-environ>=0.4,<0.11
django-redis>=4.12,<5.4
pysolr>=3.9,<3.10
redis>=3.5,<5.1
requests>=2.25,<3
uWSGI>=2.0.19
whitenoise>=5.2,<6.6
django-simple-captcha==0.6.0

# Linting
pylint>=2.17.4
pylint-django>=2.5.3
black>=23.3.0

# Helpers
pyprof2calltree>=1.4,<1.5
ipdb>=0.13,<0.14
ipython>=7.12,<9

# Country data
pycountry